+++ plugins/main/public/components/overview/fim/inventory/inventories/registry-values/inventory.tsx
@@ -5,14 +5,15 @@ import {
 } from '../../../../../common/data-source';
 import tableColumns from './table-columns';
 import managedFilters from './managed-filters';
-withSystemInventoryHardwareDataSource;
-import { getDashboard } from './dashboard';
-import { withSystemInventoryHardwareDataSource } from '../../../../it-hygiene/common/hocs/validate-system-inventory-index-pattern';
 import { withFIMRegistryValuesDataSource } from '../../../common/hocs/validate-fim-states-index-pattern';
 import { InventoryDashboardTable } from '../../../../../common/dashboards';
-import { WAZUH_SAMPLE_FILE_INTEGRITY_MONITORING } from '../../../../../../../common/constants';
-import { compose } from 'redux';
+import {
+  WAZUH_SAMPLE_FILE_INTEGRITY_MONITORING,
+  FIM_REGISTRY_VALUES_INVENTORY_ID,
+  FIM_REGISTRY_VALUES_AGENT_INVENTORY_ID,
+} from '../../../../../../../common/constants';
 import { withAgent } from '../../../../../common/hocs/with-agent';
+import { compose } from 'redux';
 
 export const InventoryFIMRegistryValues = compose(
   withAgent,
@@ -27,7 +28,12 @@ export const InventoryFIMRegistryValues = compose(
         }
         tableDefaultColumns={tableColumns}
         managedFilters={managedFilters}
-        getDashboardPanels={getDashboard}
+        getDashboardPanels={[
+          {
+            dashboardId: FIM_REGISTRY_VALUES_INVENTORY_ID,
+            agentDashboardId: FIM_REGISTRY_VALUES_AGENT_INVENTORY_ID,
+          },
+        ]}
         tableId='fim-registry-values-inventory'
         categoriesSampleData={[WAZUH_SAMPLE_FILE_INTEGRITY_MONITORING]}
       />